improve styling; add composer placeholder

+++ apps/platform/src/modules/composer/components/composer-files.tsx
@@ -7,18 +7,18 @@ export const ComposerFiles = () => {
 	const { files, removeFile } = useFileStagingContext();
 	const { openFileDetailsDialog } = useFileStagingContext();
 
+	if (files.length === 0) return null;
+
 	return (
-		<div>
-			<div className="flex flex-row overflow-x-auto gap-2">
-				{files.map((file, index) => (
-					<ComposerFile
-						file={file}
-						onRemove={() => removeFile(index)}
-						onClick={() => openFileDetailsDialog(index)}
-						key={index}
-					/>
-				))}
-			</div>
+		<div className="flex flex-row overflow-x-auto gap-2">
+			{files.map((file, index) => (
+				<ComposerFile
+					file={file}
+					onRemove={() => removeFile(index)}
+					onClick={() => openFileDetailsDialog(index)}
+					key={index}
+				/>
+			))}
 		</div>
 	);
 };
